requirements.txt
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
python-chess==1.999
numpy>=1.26.0
